agents/ingestion_agent.py
# agents/ingestion_agent.py
import pdfplumber
import tempfile
import os
from bs4 import BeautifulSoup

def ingest_pdf_bytes(pdf_bytes: bytes) -> str:
    with tempfile.NamedTemporaryFile(delete=False, suffix=".pdf") as tmp:
        tmp.write(pdf_bytes)
        path = tmp.name

    text = ""
    with pdfplumber.open(path) as pdf:
        for page in pdf.pages:
            text += page.extract_text() or ""

    os.remove(path)
    return text

def ingest_html(html: str) -> str:
    soup = BeautifulSoup(html, "html.parser")
    return soup.get_text(separator=" ")
