requirements.txt
# ----------------------------
# Core Python
# ----------------------------
python-dotenv>=1.0.0

# ----------------------------
# Web & Resource Ingestion
# ----------------------------
requests>=2.31.0
beautifulsoup4>=4.12.2
pdfplumber>=0.10.3

# ----------------------------
# NLP & Text Processing
# ----------------------------
nltk>=3.8.1
spacy>=3.7.2

# ----------------------------
# Transformers & Embeddings
# ----------------------------
sentence-transformers>=2.2.2
transformers>=4.40.0
torch>=2.1.0

# ----------------------------
# Vector Database
# ----------------------------
faiss-cpu>=1.7.4

# ----------------------------
# Reranking (Cross-Encoders)
# ----------------------------
scikit-learn>=1.3.2

# ----------------------------
# LLM & Agent Frameworks
# ----------------------------
langchain>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.4
langgraph>=0.2.0

# ----------------------------
# Dashboard / UI
# ----------------------------
streamlit>=1.33.0

# ----------------------------
# Evaluation & Utilities
# ----------------------------
numpy>=1.26.0
tqdm>=4.66.1
